chore: compile everything for Node.js 6 compat
