+++ packages/api/package.json
@@ -2,6 +2,12 @@
   "name": "@patternplate/api",
   "version": "2.0.0-16",
   "description": "Patternplate JSON API",
+  "scripts": {
+    "start": "yarn build -w",
+    "build": "babel -D src -d lib --source-maps --ignore **/*.test.js",
+    "clean": "rimraf lib",
+    "watch": "yarn build -w"
+  },
   "xo": false,
   "repository": {
     "type": "git",
@@ -22,6 +28,16 @@
   "engines": {
     "node": ">= 4",
     "npm": ">= 3"
+  },
+  "babel": {
+    "presets": [
+      "module:@patternplate/babel-preset"
+    ]
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.0.0-beta.40",
+    "@patternplate/babel-preset": "^2.0.0-16",
+    "rimraf": "^2.6.2"
   },
   "dependencies": {
     "@marionebl/sander": "^0.6.1",
