+++ packages/load-config/index.js
@@ -1,22 +1,1 @@
-const cosmiconfig = require("cosmiconfig");
-
-const DEFAULTS = {
-  docs: ["docs/**/*.md"],
-  entry: ["lib/**/demo.js"],
-  render: "@patternplate/render-default/render",
-  mount: "@patternplate/render-default/mount"
-};
-
-module.exports = async ({ cwd }) => {
-  const explorer = cosmiconfig("patternplate");
-  const config = await explorer.load(cwd);
-
-  if (!config) {
-    return {
-      config: DEFAULTS,
-      filepath: null
-    };
-  }
-
-  return config;
-};
+module.exports = require("./lib/load-config");
