+++ packages/load-config/package.json
@@ -2,6 +2,12 @@
   "name": "@patternplate/load-config",
   "version": "2.0.0-16",
   "description": "Load patternplate config",
+  "scripts": {
+    "start": "yarn build -w",
+    "build": "babel -D src -d lib --source-maps --ignore **/*.test.js",
+    "clean": "rimraf lib",
+    "watch": "yarn build -w"
+  },
   "xo": false,
   "repository": {
     "type": "git",
@@ -21,5 +27,15 @@
   },
   "dependencies": {
     "cosmiconfig": "^3.1.0"
+  },
+  "babel": {
+    "presets": [
+      "module:@patternplate/babel-preset"
+    ]
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.0.0-beta.40",
+    "@patternplate/babel-preset": "^2.0.0-16",
+    "fixturez": "^1.1.0"
   }
 }
