+++ packages/load-docs/package.json
@@ -2,6 +2,12 @@
   "name": "@patternplate/load-docs",
   "version": "2.0.0-16",
   "description": "Load patternplate docs",
+  "scripts": {
+    "start": "yarn build -w",
+    "build": "babel -D src -d lib --source-maps --ignore **/*.test.js",
+    "clean": "rimraf lib",
+    "watch": "yarn build -w"
+  },
   "xo": false,
   "repository": {
     "type": "git",
@@ -29,5 +35,15 @@
     "path-exists": "^3.0.0",
     "remark": "^8.0.0",
     "unist-util-find": "^1.0.1"
+  },
+  "babel": {
+    "presets": [
+      "module:@patternplate/babel-preset"
+    ]
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.0.0-beta.40",
+    "@patternplate/babel-preset": "^2.0.0-16",
+    "fixturez": "^1.1.0"
   }
 }
