+++ packages/pack/package.json
@@ -2,6 +2,12 @@
   "name": "@patternplate/pack",
   "version": "2.0.0-16",
   "description": "Bundle JS sources with webpack",
+  "scripts": {
+    "start": "yarn build -w",
+    "build": "babel -D src -d lib --source-maps --ignore **/*.test.js",
+    "clean": "rimraf lib",
+    "watch": "yarn build -w"
+  },
   "xo": false,
   "repository": {
     "type": "git",
@@ -22,5 +28,15 @@
   "dependencies": {
     "memory-fs": "^0.4.1",
     "webpack": "^4.0.0"
+  },
+  "babel": {
+    "presets": [
+      "module:@patternplate/babel-preset"
+    ]
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.0.0-beta.40",
+    "@patternplate/babel-preset": "^2.0.0-16",
+    "fixturez": "^1.1.0"
   }
 }
