+++ packages/search/src/query.js
