+++ packages/search/src/registry.js
