+++ packages/search/src/search.js
