+++ packages/search/src/term.js
