+++ packages/webpack-entry/index.js
@@ -1,22 +1,1 @@
-const querystring = require("querystring");
-
-const LOADER = require.resolve("./loader");
-
-module.exports = webpackEntry;
-module.exports.sync = webpackEntrySync;
-
-async function webpackEntry(entry, { cwd = process.cwd() } = {}) {
-  return `${LOADER}?${querystring.stringify({
-    entry,
-    cwd
-  })}!`;
-}
-
-function webpackEntrySync(entry, { cwd = process.cwd() } = {}) {
-  return `${LOADER}?${
-    querystring.stringify({
-      entry,
-      cwd
-    })
-  }!`;
-}
+module.exports = require("./lib/webpack-entry");
