+++ packages/webpack-entry/package.json
@@ -2,6 +2,12 @@
   "name": "@patternplate/webpack-entry",
   "version": "2.0.0-16",
   "description": "Get webpack bundle entries suitable for patternplate consumption",
+  "scripts": {
+    "start": "yarn build -w",
+    "build": "babel -D src -d lib --source-maps --ignore **/*.test.js",
+    "clean": "rimraf lib",
+    "watch": "yarn build -w"
+  },
   "xo": false,
   "repository": {
     "type": "git",
@@ -25,5 +31,15 @@
     "loader-utils": "^1.1.0",
     "lodash": "^4.17.4",
     "path-exists": "^3.0.0"
+  },
+  "babel": {
+    "presets": [
+      "module:@patternplate/babel-preset"
+    ]
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.0.0-beta.40",
+    "@patternplate/babel-preset": "^2.0.0-16",
+    "fixturez": "^1.1.0"
   }
 }
