+++ packages/webpack-entry/src/loader.js
