+++ packages/webpack-entry/src/webpack-entry.js
@@ -1,0 +1,22 @@
+const querystring = require("querystring");
+
+const LOADER = require.resolve("./loader");
+
+module.exports = webpackEntry;
+module.exports.sync = webpackEntrySync;
+
+async function webpackEntry(entry, { cwd = process.cwd() } = {}) {
+  return `${LOADER}?${querystring.stringify({
+    entry,
+    cwd
+  })}!`;
+}
+
+function webpackEntrySync(entry, { cwd = process.cwd() } = {}) {
+  return `${LOADER}?${
+    querystring.stringify({
+      entry,
+      cwd
+    })
+  }!`;
+}
